Training-Inferencing/requirements.txt
pandas==2.2.2
scikit-learn==1.4.2
flask==3.0.3

# Azure SDKs
azure-identity==1.15.0
azure-keyvault-secrets==4.8.0
azure-storage-blob==12.19.0